Get SMART configuration test working
